+++ app/models/student_response_rate.rb
@@ -19,7 +19,8 @@ class StudentResponseRate
   def response_count
     @response_count ||= @subcategory.measures.map do |measure|
       measure.student_survey_items.map do |survey_item|
-        survey_item.survey_item_responses.where(school: @school, academic_year: @academic_year).count
+        survey_item.survey_item_responses.where(school: @school,
+                                                academic_year: @academic_year).exclude_boston.count
       end.sum
     end.sum
   end