+++ db/migrate/20220510232626_add_foreign_key_between_schools_and_districts.rb
@@ -1,0 +1,5 @@
+class AddForeignKeyBetweenSchoolsAndDistricts < ActiveRecord::Migration[7.0]
+  def change
+    add_foreign_key :schools, :districts
+  end
+end