+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_05_03_180142) do
+ActiveRecord::Schema[7.0].define(version: 2022_05_10_232626) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_stat_statements"
   enable_extension "plpgsql"
@@ -381,6 +381,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_05_03_180142) do
   add_foreign_key "respondents", "academic_years"
   add_foreign_key "respondents", "schools"
   add_foreign_key "scales", "measures"
+  add_foreign_key "schools", "districts"
   add_foreign_key "subcategories", "categories"
   add_foreign_key "survey_item_responses", "academic_years"
   add_foreign_key "survey_item_responses", "schools"